% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_RNAVelocity.R
\name{getExtrapolatedStateInDCSpace}
\alias{getExtrapolatedStateInDCSpace}
\title{Transform the extrapolated space from RNA velocity into RNA velocity.}
\usage{
getExtrapolatedStateInDCSpace(dataList)
}
\arguments{
\item{dataList}{A Revelio object that contains a rotation matrix and a velocity matrix.}
}
\value{
Returns the same Revelio object given as input but additionally with the coordinates of the extrapolated state in DC space.
}
\description{
'getExtrapolatedStateInDCSpace' calculates the extrapolated state of each cells and transforms it into the coordinate system of the dynamical components.
}
\details{
For each cell the extrapolated state is calculated (using the RNA velocity matrix calculated previously) via the model s_ij=s_ij+v. This extrapolated state is then transformed into the dynamical component space using the same transformation matrices generated by only the exonic data.
}
