% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_Rotation.R
\name{getOptimalRotation}
\alias{getOptimalRotation}
\title{Get Optimal Rotation.}
\usage{
getOptimalRotation(dataList, boolPlotResults = FALSE)
}
\arguments{
\item{dataList}{A Revelio object that contains a raw data matrix, assigned cell cycle phases and PCA information.}

\item{boolPlotResults}{TRUE/FALSE if resulting cell cycle should be plotted.}
}
\value{
Returns the same Revelio object given as input with an added panel dc in the transformed data panel. Also the angle, radius und pseudotime ordering is added to the cellInfo panel.
}
\description{
'getOptimalRotation' rotates the PCA space to reveal the cell cycle in the first two dimensions.
}
\details{
The PCs that are dominated by cell cycle effects are isolated and sequences of three-dimensional rotations are performed, minimizing the cluster score in the corresponding third dimension. This isolates the cell cycle signal into the first dimensions.
}
