% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_PCA.R
\name{getPCAData}
\alias{getPCAData}
\title{Perform PCA.}
\usage{
getPCAData(dataList)
}
\arguments{
\item{dataList}{A Revelio object that contains a raw data matrix assigned cell cycle phases.}
}
\value{
Returns the same Revelio object given as input but now with added PCA information under the transformedData panel.
}
\description{
'getPCAData' performs a principal component analysis (PCA).
}
\details{
The data is transformed to z-scores and PCA is applied. When creating the Revelio object, the user should have defined which genes should be utilized for the PCA. By default this should be 'variableGenes' but can be changed to 'allGenes'. Be aware that the algorithm takes much longer when using all genes.
}
