% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_RNAVelocity.R
\name{getVelocityGridAlongThirdDimensions}
\alias{getVelocityGridAlongThirdDimensions}
\title{Calculate the velocity grid along a third dimension.}
\usage{
getVelocityGridAlongThirdDimensions(
  dataList,
  whichThirdDimension = 3,
  sigma = 1,
  numberOfGridPointsOneSide = 20,
  boolPlotResults = FALSE
)
}
\arguments{
\item{dataList}{A Revelio object that contains information about the extrapolated state of cells.}

\item{whichThirdDimension}{The number of the dimension to be investigated. Needs to be an integer.}

\item{sigma}{The smoothing parameter for the Gaussian kernel.}

\item{numberOfGridPointsOneSide}{The number of grid points along one side. The total grid will be a numberOfGridPointsOneSide-by-numberOfGridPointsOneSide grid.}

\item{boolPlotResults}{TRUE/FALSE if the velocity grid plot should be shown.}
}
\value{
Returns the same Revelio object given as input with the additional information about the RNA velocity at grid points.
}
\description{
'getVelocityGridAlongThirdDimensions' places a regular grid on the two-dimensional plot consisting of cell cycle pseudotime and a higher order DC and calculates the average velocity for each grid point as described in La Manno et al. 2018 ('RNA velocity of single cells').
}
\details{
A grid is generated overlaying the two-dimensional plot of cell cycle pseudotime and a higher order DC. For each grid point, the RNA velocity is calculated by using a Gaussian kernel as explained in La Manno et al. 2018 ('RNA velocity of single cells').
}
