% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_RNAVelocity.R
\name{getVelocityMatrix}
\alias{getVelocityMatrix}
\title{Get Velocity Matrix.}
\usage{
getVelocityMatrix(dataList, rawIntronicData)
}
\arguments{
\item{dataList}{A Revelio object that contains a raw data matrix with assigned cell cycle phases.}

\item{rawIntronicData}{A genes-by-cells matrix of raw intronic sequencing data. Rows should contain gene names, columns should contain cell IDs, entries of the matrix are integer numbers. The cell IDs should coincide with the cell IDs provided in the exonic matrix previously.}
}
\value{
Returns the same Revelio object given as input but now with an added intronic data matrix and the velocity matrix.
}
\description{
'getVelocityMatrix' reads the intronic data, filters the existing gene list and calculates the RNA velocity matrix according to Model I in La Manno et al. 2018 ('RNA velocity of single cells').
}
\details{
The intronic data should be corresponding to the exonic data previously provided. First, the offset parameter is estimated, then the gamma parameter and finally the velocity matrix from the model I from La Manno et al. 2008 ('RNA velocity of single cells') where the time derivative of spliced counts is assumed constant.
}
