% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_RNAVelocity.R
\name{plotVelocityOnAveragedTrajectory}
\alias{plotVelocityOnAveragedTrajectory}
\title{Calculate the velocity along an averaged trajectory.}
\usage{
plotVelocityOnAveragedTrajectory(dataList, numberOfAnchors = 10)
}
\arguments{
\item{dataList}{A Revelio object that contains information about the extrapolated state of cells.}

\item{numberOfAnchors}{The number of anchor points used for the spline of the averaged trajectory. This also defines the points at which RNA velocity will be calculated.}
}
\value{
Returns the same Revelio object given as input.
}
\description{
'getVelocityOnAveragedTrajectory' places an average trajectory through the data cloud and calculates the average velocity for each anchor point as described in La Manno et al. 2018 ('RNA velocity of single cells').
}
\details{
A trajectory is placed through the data cloud as a spline in polar coordinates with a number of anchor points specified by the user. For each anchor point, the RNA velocity is calculated by using a Gaussian kernel as explained in La Manno et al. 2018 ('RNA velocity of single cells').
}
