% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revelioTestData_cyclicGenes}
\alias{revelioTestData_cyclicGenes}
\title{A table of marker gene lists.}
\format{
A data frame with 216 rows (length of the longest marker gene list) and 5 variables (cell cycle phases and transitions: G1.S, S, G2, G2.M, M.G1).
}
\usage{
revelioTestData_cyclicGenes
}
\description{
A dataset that contains multiple lists of marker genes corresponding to cell cycle phases or cell cycle phase transitions.
}
\keyword{datasets}
