% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revelioTestData_rawDataMatrix}
\alias{revelioTestData_rawDataMatrix}
\title{A single-cell gene expression data matrix.}
\format{
A data frame with 22717 rows (genes) and 1477 variables (cells).
}
\usage{
revelioTestData_rawDataMatrix
}
\description{
A dataset that contains a gene-by-cell matrix with UMI counts as entries.
}
\keyword{datasets}
