% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{revelioTestData_rawIntronicMatrix}
\alias{revelioTestData_rawIntronicMatrix}
\title{A single-cell gene expression data matrix of intronic counts.}
\format{
A data frame with 19201 rows (genes) and 1477 variables (cells).
}
\usage{
revelioTestData_rawIntronicMatrix
}
\description{
A dataset that contains a gene-by-cell matrix with UMI counts as entries.
}
\keyword{datasets}
