% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/functions_InitializationCCPhaseAssignFiltering.R
\name{createRevelioObject}
\alias{createRevelioObject}
\title{Create Revelio Object.}
\usage{
createRevelioObject(
  datasetID = "data1",
  rawData = NULL,
  cyclicGenes = NULL,
  cellType = "unknown",
  sequTechnique = "unknown",
  dataPath = "unknown",
  lowernGeneCutoff = 0,
  uppernUMICutoff = 10^7,
  ccPhaseAssignBasedOnIndividualBatches = TRUE,
  ccPhaseAssignGeneBucketListFileLocation = "unknown",
  ccPhaseAssignGeneBucketListSheetIndex = 1,
  ccPhaseAssignThresholdForCorSingleGeneToAvgExpression = 0.2,
  ccPhaseAssignMaxDiffOfCCPhasesBetweenHighestScores = 1,
  ccPhaseAssignThresholdHighestPhaseScore = 0.75,
  ccPhaseAssignThresholdSecondHighestPhaseScore = 0.5,
  pcaGenes = "variableGenes",
  thresholdsForPCWeightSignificance = 0.05,
  ccDurationG1 = 0.487 * 19.33,
  ccDurationS = 0.392 * 19.33,
  ccDurationG2 = 0.093 * 19.33,
  ccDurationM = 0.028 * 19.33,
  ccDurationTotal = 19.33
)
}
\arguments{
\item{datasetID}{ID should correspond to ID provided in supplementary file.}
}
\value{
Returns Revelio object.
}
\description{
'createRevelioObject' returns a Revelio object according to the specified parameters.
}
\details{
testtest
}
